requirements.txt
requests
xprintidle
click
pillow
